Checkpoint before follow-up message
