+++ src/feature_selection/vectorbt_extensions/vectorbt_unified_framework.py
@@ -37,9 +37,9 @@ from src.utils.hardware import (
     memory_efficient,
     performance_tracked,
     smart_cache,
-    WorkloadType,
-    OptimizationLevel
+    WorkloadType
 )
+from src.utils.hardware.constants import OptimizationLevel
 
 # Import utilities
 from src.utils.tprint import tprint, tprint_success, tprint_warning, tprint_performance, tprint_debug
